Done with the first test. Uploading the test on a server.

+++ simple_flows/simple_flow_generator/flow_dir/sock_addr_struct.py
@@ -6,14 +6,18 @@ easier to understand this file and extedn it in the future if Linux changes its 
 For more reference please refer to the Linux counterparts.
 '''
 
-from ctypes import Structure, c_ulong, c_short, c_ushort, c_char
+from ctypes import Structure
+from ctypes import c_uint16
+from ctypes import c_ushort
+from ctypes import c_char
+from ctypes import c_uint32
 
 
 '''
 The Internet Address class stores an IPv4 address.
 '''
 class In_Address(Structure):
-    _fields_ = [("s_addr", c_ulong)]
+    _fields_ = [("s_addr", c_uint32)]
 
 
 '''
@@ -21,5 +25,5 @@ The Socket Address Interet class stores values that its Linux
 counterpart stores.
 '''
 class Sockaddr_In(Structure):
-    _fields_ = [("sin_family", c_short), ("sin_port", c_ushort), ("sin_addr", In_Address), ("sin_zero", (c_char*8))]
+    _fields_ = [("sin_family", c_ushort), ("sin_port", c_uint16), ("sin_addr", In_Address), ("sin_zero", (c_char*8))]
